requirements.txt
beautifulsoup4==4.12.3
blinker==1.7.0
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
contourpy==1.2.0
csv23==0.3.4
cycler==0.12.1
DateTime==5.4
emoji==2.12.1
et-xmlfile==1.1.0
Flask==3.0.1
fonttools==4.48.1
idna==3.6
insta-scrape==2.1.2
instaloader==4.13.1
itsdangerous==2.1.2
Jinja2==3.1.3
joblib==1.3.2
kiwisolver==1.4.5
MarkupSafe==2.1.4
matplotlib==3.8.3
nltk==3.8.1
numpy==1.26.3
openpyxl==3.1.5
packaging==23.2
pandas==2.2.0
pillow==10.2.0
pyparsing==3.1.1
python-dateutil==2.8.2
python-dotenv==1.0.1
pytz==2023.3.post1
regex==2023.12.25
requests==2.31.0
scikit-learn==1.4.0
scipy==1.12.0
six==1.16.0
soupsieve==2.5
threadpoolctl==3.3.0
tqdm==4.66.2
typing_extensions==4.12.2
tzdata==2023.4
urllib3==2.1.0
Werkzeug==3.0.1
wordcloud==1.9.3
zope.interface==6.1
